% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{change_path}
\alias{change_path}
\title{Change the path for a \code{tnseq} directory.}
\usage{
change_path(tnseq, newpath)
}
\arguments{
\item{tnseq}{A \code{tnseq} object.}

\item{newpath}{String with location of new directory.}
}
\value{
The modified \code{tnseq} object.
}
\description{
Change the path for a \code{tnseq} directory.
}

