% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{create_tnseq_directory}
\alias{create_tnseq_directory}
\title{Create empty Tn-seq directory with sample files.}
\usage{
create_tnseq_directory(path)
}
\arguments{
\item{path}{Directory to be created.}
}
\description{
Create empty Tn-seq directory with sample files.
}
\seealso{
\code{\link{create_tnseq_experiment}}
}

