% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{deriv_beta2}
\alias{deriv_beta2}
\title{obtain standardard deviation for the difference between the mean for
each of the normal genes with the common mean for pseudogenes
covariate is included, the dimension of coefficient regression is 1.
the estimated second derivative of loglikelihood w.r.t beta}
\usage{
deriv_beta2(betaest, muest, nbet = 1, covariance = TRUE)
}
\arguments{
\item{betaest}{this reads like beta test to me}

\item{muest}{this is the muest mu ever!}

\item{nbet}{nbets are yummy}

\item{covariance}{is this a covariance model?}
}
\value{
stuff
}
\description{
ok
}

