% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{deriv_betamu}
\alias{deriv_betamu}
\title{the estimated second derivative of loglikelihood w.r.t beta and mu}
\usage{
deriv_betamu(betaest, muest, covariance = TRUE)
}
\arguments{
\item{betaest}{more betas please}

\item{muest}{and more mues}

\item{covariance}{more covariances also}
}
\value{
stuff
}
\description{
hmm ok
}

