% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{deriv_mu2}
\alias{deriv_mu2}
\title{the estimated second derivative of loglikelihood w.r.t mu}
\usage{
deriv_mu2(dat, regression_estimate_coef, mean_count_estimate)
}
\arguments{
\item{dat}{the data frame}

\item{regression_estimate_coef}{the coefficients of the regression}

\item{mean_count_estimate}{guess!}
}
\value{
stuff
}
\description{
the estimated second derivative of loglikelihood w.r.t mu
}

