% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{deriv_SD_diff}
\alias{deriv_SD_diff}
\title{derive the variance of the difference b.w. between the mean for each of
the normal genes with common mean for pseudogenes}
\usage{
deriv_SD_diff(bet, mu)
}
\arguments{
\item{bet}{almost like bert}

\item{mu}{and nu}
}
\value{
stuff
}
\description{
I agree!
}

