% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{deriv_SD_mu}
\alias{deriv_SD_mu}
\title{derive the variance for the mean offor each of
the normal genes}
\usage{
deriv_SD_mu(bet, mu, covariance = TRUE)
}
\arguments{
\item{bet}{is this bert?}

\item{mu}{and ernie!?}

\item{covariance}{covar?}
}
\description{
hmm ok
}

