% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsEM.r
\name{DeriveMeanRegression}
\alias{DeriveMeanRegression}
\title{get a mean regression from a tnseq object
Once an EM has been performed, find the mean of the regressions?}
\usage{
DeriveMeanRegression(me, nbet = 1)
}
\arguments{
\item{me}{the tnseq object including the final regression state}

\item{nbet}{I suspect this is related to the difference between a covariance model and not
but as it stands it looks kind of dumb to me}
}
\description{
get a mean regression from a tnseq object
Once an EM has been performed, find the mean of the regressions?
}

