% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsEM.r
\name{DeriveMeanSD}
\alias{DeriveMeanSD}
\title{extract the standard deviation of a set of means using the evil nbets
yet another function which makes 0 sense to me}
\usage{
DeriveMeanSD(me, nbet = 1)
}
\arguments{
\item{me}{the tnseq object}

\item{nbet}{rawr1}
}
\value{
the standard deviations
}
\description{
extract the standard deviation of a set of means using the evil nbets
yet another function which makes 0 sense to me
}

