% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsEM.r
\name{DeriveMeanSquared}
\alias{DeriveMeanSquared}
\title{Similar to DeriveRegressionSquared I do not understand what is going on here, either}
\usage{
DeriveMeanSquared(me)
}
\arguments{
\item{me}{the tnseq object with the final regression state and mean counts}
}
\value{
a new set of means squared using fancy math
}
\description{
Similar to DeriveRegressionSquared I do not understand what is going on here, either
}

