% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsEM.r
\name{DeriveRegressionSquared}
\alias{DeriveRegressionSquared}
\title{ok I don't understand what is the point of -derivp1 ^ 2
This is one of a shrinking majority of functions in this code for which I have no understanding}
\usage{
DeriveRegressionSquared(me)
}
\arguments{
\item{me}{the tnseq object including the final regression state}
}
\value{
the sum of derivp and derivp2 of course, gosh
}
\description{
ok I don't understand what is the point of -derivp1 ^ 2
This is one of a shrinking majority of functions in this code for which I have no understanding
}

