% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{em.algo}
\alias{em.algo}
\title{the main function of EM Algorithm}
\usage{
em.algo(dat, initial_mean_counts, initial_regression, maxit = 600,
  tolerance = 1e-06)
}
\arguments{
\item{dat}{the data frame}

\item{initial_mean_counts}{this really doesn't belong here I think}

\item{initial_regression}{neither does this}

\item{maxit}{maximum number of iterations before it drops out}

\item{tolerance}{minimum difference between two iterations before
it pops out and returns}
}
\value{
stuff
}
\description{
the main function of EM Algorithm
}

