% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{em_algorithm}
\alias{em_algorithm}
\title{the main function of EM Algorithm}
\usage{
em_algorithm(mu_inits, beta_inits, maxit = 500, tolerance = 1e-06,
  covariance = FALSE)
}
\arguments{
\item{mu_inits}{mu inits are awesome!}

\item{beta_inits}{I want to init some betas}

\item{maxit}{500 iterations sound good to me}

\item{tolerance}{1 in a million is awesome}

\item{covariance}{is this a covariance model?}
}
\value{
stuff
}
\description{
This does the actual work, and for a moment I thought I understood what was happening, but that was woefully not true.
}

