% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{estep}
\alias{estep}
\title{Expectation Step in the EM algorithm}
\usage{
estep(count, mean_counts_G, regression)
}
\arguments{
\item{mean_counts_G}{dunno}

\item{regression}{the glm regression}

\item{counts}{our count list}
}
\value{
z_temp which is 0 or an inverse_logit
}
\description{
Expectation Step in the EM algorithm
}

