% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsEM.r
\name{EstimateMeans}
\alias{EstimateMeans}
\title{make a list of the normal genes and put into it the estimated (means/position)/gene}
\usage{
EstimateMeans(me)
}
\arguments{
\item{me}{the tnseq object containing counts/position, the number of normal genes, and gene types}
}
\value{
estimated means, probably should be changed to me
}
\description{
make a list of the normal genes and put into it the estimated (means/position)/gene
}

