% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{get_path}
\alias{get_path}
\title{Return paths for experiment files and directories.}
\usage{
get_path(tnseq, file = NA, dir = NA)
}
\arguments{
\item{tnseq}{A \code{tnseq} object.}

\item{file}{Optional filename to append to end of path.}

\item{dir}{Subdirectory (\code{input}, \code{log}, etc.).  If not given,
   returns the path to the \code{tnseq} directory.}
}
\value{
String representing the directory or file path.  Directory paths
   will always include a trailing "/".
}
\description{
Return paths for experiment files and directories.
}
\examples{
get_path(tnseq, dir="split")
get_path(tnseq, file="lane1.fastq", dir="input")
}

