% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{initialize_means}
\alias{initialize_means}
\title{initialize the list of means}
\usage{
initialize_means(mean_estimates, df, in_sites = NULL)
}
\arguments{
\item{mean_estimates}{a data frame of counts}

\item{df}{a data frame of counts}

\item{in_sites}{the list of insertion sites}
}
\value{
a length data structure
}
\description{
initialize the list of means
}

