% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunEM.r
\name{InitializeEM}
\alias{InitializeEM}
\title{Given a tnseq object containing counts by positions and some annotation information
add to the tnseq object the number of normal genes, the number of insertion sites,
the number of pseudogene positions, and some initial estimates}
\usage{
InitializeEM(me, cutoff = 5)
}
\arguments{
\item{me}{the original tnseq object}

\item{cutoff}{an arbitrary 5 which I don't yet understand}
}
\value{
a bigger tnseq object
}
\description{
Given a tnseq object containing counts by positions and some annotation information
add to the tnseq object the number of normal genes, the number of insertion sites,
the number of pseudogene positions, and some initial estimates
}

