% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{insertion_sites}
\alias{insertion_sites}
\title{calculate the number of insertion sites}
\usage{
insertion_sites(df)
}
\arguments{
\item{df}{a data frame of counts}
}
\value{
a length data structure
}
\description{
calculate the number of insertion sites
}

