% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TnSeqSetup.r
\name{InsertionSites}
\alias{InsertionSites}
\title{Given a set of hits/position and gene names
figure out how many possible insertion sites exist and return them in a very peculiar fashion}
\usage{
InsertionSites(me)
}
\arguments{
\item{me}{a tnseq object containing the data frame of interest
I should probably have this return me}
}
\value{
the na.omitted set of sites
}
\description{
Given a set of hits/position and gene names
figure out how many possible insertion sites exist and return them in a very peculiar fashion
}

