% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{inverse_logit}
\alias{inverse_logit}
\title{Do an inverse logit}
\usage{
inverse_logit(number, derivative = 0)
}
\arguments{
\item{number}{a number}

\item{derivative}{0 means calculate the logit, 1 or 2 a 1st/2nd derivatives}
}
\value{
another number
}
\description{
Do an inverse logit
}

