% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsEM.r
\name{InverseLogit}
\alias{InverseLogit}
\title{Given a number, calculate its inverse logit, the derivative of it, or second derivative}
\usage{
InverseLogit(number, derivative = 0)
}
\arguments{
\item{number}{the number to modify}

\item{derivative}{0 means just calculate the value, 1 or 2 as for derivatives}
}
\value{
a new number
}
\description{
Given a number, calculate its inverse logit, the derivative of it, or second derivative
}

