% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{invlogit}
\alias{invlogit}
\title{Calculate the inverse logit}
\usage{
invlogit(z)
}
\arguments{
\item{z}{An input number to be inversed}
}
\value{
inv_logit
}
\description{
yeah I will admit, I don't have a clue what this does nor why
But the code calls e^(-1 * z), adds 1 to the result, then flips it to 1/result
}

