% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{invlogit_deriv}
\alias{invlogit_deriv}
\title{Calculate the inverse logit derivative}
\usage{
invlogit_deriv(z)
}
\arguments{
\item{z}{An input number to be inverse logitted}
}
\value{
an inverse derivative, I guess
}
\description{
Same as above, I don't see why this is happening
But it takes a number and subtracts from it 2*log(1+exp(z)) for a result
Then takes that and does e^result
}

