% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{invlogit_second_deriv}
\alias{invlogit_second_deriv}
\title{Take the second derivative of an inverse logit}
\usage{
invlogit_second_deriv(z)
}
\arguments{
\item{z}{a number to be second derivatived logitted}
}
\value{
an inverse second derivative, maybe
}
\description{
hmm yeah whatever
}

