% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{lik0}
\alias{lik0}
\title{A dirty likelihood function}
\usage{
lik0(p, mean_counts)
}
\arguments{
\item{p}{a number}

\item{mean_counts}{a list of counts}
}
\value{
another number
}
\description{
A dirty likelihood function
}

