% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsEM.r
\name{Likelihood}
\alias{Likelihood}
\title{a cheapo likelihood function}
\usage{
Likelihood(num, counts)
}
\arguments{
\item{num}{given a number}

\item{counts}{and some count data}
}
\value{
where does that number fit with respect to the counts using an exponent?
}
\description{
a cheapo likelihood function
}

