% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{load_tnseq_experiment}
\alias{load_tnseq_experiment}
\title{Load a \code{tnseq} object from existing directory.}
\usage{
load_tnseq_experiment(path)
}
\arguments{
\item{path}{Experimental directory to load.}
}
\value{
A \code{tnseq} object.
}
\description{
Load a \code{tnseq} object from existing directory.
}
\seealso{
\code{\link{create_tnseq_directory}}
}

