% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{make_result_df}
\alias{make_result_df}
\title{create a data frame of results from the output of the em algorithm}
\usage{
make_result_df(dat, em_output, sig_cutoff = 0.05, p_adjust = "BH")
}
\arguments{
\item{dat}{the original data}

\item{em_output}{the result from em}

\item{sig_cutoff}{maximum qvalue to accept as 'essential'}

\item{p_adjust}{method to use to adjust the pvalues}
}
\value{
a data frame of 6 columns
}
\description{
create a data frame of results from the output of the em algorithm
}

