% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{maximize_read_counts}
\alias{maximize_read_counts}
\title{Maximization Step---mu in the EM algorithm}
\usage{
maximize_read_counts(dat, count, z_temp)
}
\arguments{
\item{dat}{the data frame}

\item{count}{the count list}

\item{z_temp}{something else}
}
\value{
a maximum
}
\description{
Maximization Step---mu in the EM algorithm
}

