% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{maximize_regression_coef}
\alias{maximize_regression_coef}
\title{Maximization Step---beta in the EM algorithm}
\usage{
maximize_regression_coef(z_temp)
}
\arguments{
\item{z_temp}{a number}
}
\value{
a maximized regression
}
\description{
Maximization Step---beta in the EM algorithm
}

