% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunEM.r
\name{MaximizeReadCounts}
\alias{MaximizeReadCounts}
\title{multiply all the read counts by the current regression state and sum them,
then sum the regression states, and take the ratio for each count position}
\usage{
MaximizeReadCounts(me, regression_step)
}
\arguments{
\item{me}{a tnseq object}

\item{regression_step}{an individual iteration of the regression}
}
\value{
maximized mean counts
}
\description{
multiply all the read counts by the current regression state and sum them,
then sum the regression states, and take the ratio for each count position
}

