% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunEM.r
\name{MaximizeRegression}
\alias{MaximizeRegression}
\title{Use a general linear model on the regression and return the coefficients
I have a fairly basic understanding of glm, I am not entirely certain how
this counts as a maximization, but that is probably due to my ignorance.}
\usage{
MaximizeRegression(regression_step)
}
\arguments{
\item{regression_step}{a new regression, which is in turn a call to
InverseLogit of the sum of mean_counts_G and the last regression}
}
\value{
a maximized version of the regression step
}
\description{
Use a general linear model on the regression and return the coefficients
I have a fairly basic understanding of glm, I am not entirely certain how
this counts as a maximization, but that is probably due to my ignorance.
}

