% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{mstep_beta}
\alias{mstep_beta}
\title{Maximization Step---beta in the EM algorithm}
\usage{
mstep_beta(X, z_temp, covariance = TRUE)
}
\arguments{
\item{X}{oh noes the dredded capital x}

\item{z_temp}{sigh}

\item{covariance}{is this a cov model?}
}
\value{
the coefficients from the general linear model fit of this crap
}
\description{
oh so z_temp is a parameter for this function but not the previous!?
}

