% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{mstep_mu}
\alias{mstep_mu}
\title{perform a maximization!}
\usage{
mstep_mu(count, z_temp)
}
\arguments{
\item{count}{counts are fun}

\item{z_temp}{I am pretty sure you aren't supposed to name your variables z_temp}
}
\value{
maximized pain
}
\description{
this does something I am sure which includes unicorns and ponies
}

