% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/configure.R
\name{parallel}
\alias{get_cores}
\alias{parallel}
\alias{use_multiple_cores}
\alias{use_single_core}
\title{Parallel processing features.}
\usage{
use_multiple_cores(ncores)

use_single_core()

get_cores()
}
\description{
Tn-seq computations can be carried out in parallel on multiple cores.
Parallelization is accomplished using the "parallel" package.  The following
functions enable and configure parallel processing.
}
\section{Functions}{
\itemize{
\item \code{use_multiple_cores}: Set the number of cores.  Set to 1 to turn off
parallelization.

\item \code{use_single_core}: Alias for use_multiple_cores(1).

\item \code{get_cores}: Return the number of cores to be used.
}}

