% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{perform_tnseq_em}
\alias{perform_tnseq_em}
\title{Given a dataframe with gene names, counts by position, and a
boolean describing whether they are pseudogenes, perform the
expectation maximization algorithm}
\usage{
perform_tnseq_em(dat, covariance = FALSE)
}
\arguments{
\item{dat}{A data frame with (currently) 3 columns, gene, count,
and gene_type (pseudogene is TRUE)}

\item{covariance}{when performing the linear fits, have a
covariance?}
}
\value{
a matrix (soon to be data frame) of the data
        probably will be made into a list with other data too
}
\description{
Given a dataframe with gene names, counts by position, and a
boolean describing whether they are pseudogenes, perform the
expectation maximization algorithm
}

