% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunEM.r
\name{PerformEM}
\alias{PerformEM}
\title{Perform the Expectation Maximization algorithm given a set of mean counts and tolerance}
\usage{
PerformEM(me, maxit = 600, tolerance = 1e-06)
}
\arguments{
\item{me}{a tnseq object containing an initial set of mean counts by position}

\item{maxit}{the maximum number of iterations before giving up}

\item{tolerance}{the maximum difference between iterations before dropping out and returning an answer}
}
\value{
me but with final_mean_counts and final_regression filled out
}
\description{
Perform the Expectation Maximization algorithm given a set of mean counts and tolerance
}

