% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnseqoo.r
\name{ReadAnnotation}
\alias{ReadAnnotation}
\title{Read an annotation file or data frame and make sure all is kosher}
\usage{
ReadAnnotation(input, ...)
}
\arguments{
\item{input}{a data frame or filename of annotation information}

\item{...}{extra arguments most likely passed to read.table}
}
\value{
a data frame of annotation information
}
\description{
Read an annotation file or data frame and make sure all is kosher
}

