% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TnSeqSetup.r
\name{ReadCounts}
\alias{ReadCounts}
\title{Read a count table and make sure that it has some stuff needed to work}
\usage{
ReadCounts(input, ...)
}
\arguments{
\item{input}{a data frame of counts}

\item{...}{extra arguments most likely passed to read.table}
}
\value{
a data frame of counts by gene/position
}
\description{
Read a count table and make sure that it has some stuff needed to work
}

