% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnseqoo.r
\name{ReadGenome}
\alias{ReadGenome}
\title{Gather a genome for analyses of possible insertion sites, gene lengths, whatever}
\usage{
ReadGenome(fasta, ...)
}
\arguments{
\item{fasta}{a fast file to read}

\item{...}{extra arguments likely to FaFile but maybe something else if I think on it}
}
\value{
a fasta object I can use for PDict()
}
\description{
Gather a genome for analyses of possible insertion sites, gene lengths, whatever
}

