% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunEM.r
\name{ScoreEM}
\alias{ScoreEM}
\title{given a completed EM run, score the result, drop in some p/q values
and return a table of the results}
\usage{
ScoreEM(me, sig_cutoff = 0.05, p_adjust = "BH")
}
\arguments{
\item{me}{the finalized EM run}

\item{sig_cutoff}{maximum adjusted p value}

\item{p_adjust}{the p adjustment to use}
}
\value{
a dataframe of yes/no by gene with scores
}
\description{
given a completed EM run, score the result, drop in some p/q values
and return a table of the results
}

