% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/split.R
\name{split_inputs}
\alias{split_inputs}
\title{Trim raw reads and split by DNA barcodes.}
\usage{
split_inputs(tnseq, ...)
}
\arguments{
\item{tnseq}{A tnseq object.}
}
\value{
A tnseq object with updated filelog.
}
\description{
Adapters are trimmed, and barcodes are used to split input FASTQ files into
collapsed FASTA files (in the split/ directory).
}

