% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TnSeq.r
\name{TnSeq}
\alias{TnSeq}
\title{Create a TnSeq object given the camelcaseish nature of the original source
code, perhaps I will finally start following the google/etc code style guide.
The code from which I am starting provides a position sorted data frame with
gene names as a field, this is a bit peculiar and I will likely take some
liberties with it, especially given that other code I have written is able
to figure out the number of possible insertion sites denovo.}
\usage{
TnSeq(count_table, annotation = NULL, genome = NULL, task = "em_nocov",
  ...)
}
\arguments{
\item{count_table}{either a data frame or filename with counts by position}

\item{annotation}{a data frame or filename with annotation information}

\item{genome}{a filename containing a fasta genome}
}
\value{
a TnSeq object
}
\description{
This begs a question of Yoann, should I assume any other insertion sites
than TA?  All tnseq work I have read uses mariner, but that isn't necessarily
the only transposable element one works with?
}

