% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{change_path}
\alias{change_path}
\title{Change the path for a \code{tnseq} directory.}
\usage{
change_path(tnseq, ...)
}
\arguments{
\item{tnseq}{A \code{tnseq} object.}

\item{newpath}{String with location of new directory.}
}
\value{
The modified \code{tnseq} object.

A \code{tnseq} object stores the path given in
\code{\link{create_tnseq_experiment}} to locate experiment files.  When
renaming a tnseq directory, use this command to update the \code{tnseq}
object.
}
\description{
Change the path for a \code{tnseq} directory.
}
\seealso{
\code{\link{get_path}}
}

