% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{deriv_betamu}
\alias{deriv_betamu}
\title{the estimated second derivative of loglikelihood w.r.t beta and mu}
\usage{
deriv_betamu(dat, regression_estimate_coef, mean_count_estimate, nbet = 1)
}
\arguments{
\item{dat}{the data frame}

\item{regression_estimate_coef}{bingo!}

\item{mean_count_estimate}{the mean counts}

\item{nbet}{I suspect this has something to do with the covariance
model or lack thereof}
}
\value{
stuff
}
\description{
the estimated second derivative of loglikelihood w.r.t beta and mu
}

