% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{deriv_mu2}
\alias{deriv_mu2}
\title{the estimated second derivative of loglikelihood w.r.t mu}
\usage{
deriv_mu2(betaest, muest, covariance = TRUE)
}
\arguments{
\item{betaest}{more betas}

\item{muest}{more mus!}

\item{covariance}{is this a covariance model?}
}
\value{
stuff
}
\description{
yep I agree
}

