% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{deriv_SD_diff}
\alias{deriv_SD_diff}
\title{derive the variance of the difference b.w. between the mean for each of
the normal genes with common mean for pseudogenes}
\usage{
deriv_SD_diff(dat, regression_coef, mean_counts, nbet = 1)
}
\arguments{
\item{dat}{my data}

\item{regression_coef}{the regression coefficients}

\item{mean_counts}{the counts}

\item{nbet}{the nbetas or something}
}
\value{
stuff
}
\description{
derive the variance of the difference b.w. between the mean for each of
the normal genes with common mean for pseudogenes
}

