% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximization_tnseq.r
\name{deriv_SD_mu}
\alias{deriv_SD_mu}
\title{derive the variance for the mean offor each of
the normal genes}
\usage{
deriv_SD_mu(dat, regression_coef, mean_counts, nbet = 1)
}
\arguments{
\item{dat}{the data}

\item{regression_coef}{the regression coefficients}

\item{mean_counts}{the counts}

\item{nbet}{the nbets}
}
\value{
stuff
}
\description{
derive the variance for the mean offor each of
the normal genes
}

