% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{estep}
\alias{estep}
\title{Expectation Step in the EM algorithm}
\usage{
estep(count, muG, beta, X, covariance = TRUE)
}
\arguments{
\item{count}{counts!}

\item{muG}{oh come on really}

\item{beta}{I am getting no love}

\item{X}{oh a capital x, I am sure that makes sense}

\item{covariance}{is this a cov model?}
}
\value{
z_temp, because I said so
}
\description{
do some stuff, how the hell did z_temp get here?
}

