% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.r
\name{lik0}
\alias{lik0}
\title{wtf is lik0 supposed to mean}
\usage{
lik0(p, mu)
}
\arguments{
\item{p}{a vector}

\item{mu}{another vector}
}
\value{
lik, no lik0 man
}
\description{
Given 2 vectors of the same length, add the first to
((1-itself) * e^(-1 * second vector))
I am sure this makes sense
}

